+++ braintree/customer_gateway.py
@@ -6,7 +6,6 @@ from braintree.ids_search import IdsSearch
 from braintree.resource import Resource
 from braintree.resource_collection import ResourceCollection
 from braintree.successful_result import SuccessfulResult
-from braintree.transparent_redirect import TransparentRedirect
 
 class CustomerGateway(object):
     def __init__(self, gateway):
@@ -16,10 +15,6 @@ class CustomerGateway(object):
     def all(self):
         response = self.config.http().post(self.config.base_merchant_path() + "/customers/advanced_search_ids")
         return ResourceCollection({}, response, self.__fetch)
-
-    def confirm_transparent_redirect(self, query_string):
-        id = self.gateway.transparent_redirect._parse_and_validate_query_string(query_string)["id"][0]
-        return self._post("/customers/all/confirm_transparent_redirect_request", {"id": id})
 
     def create(self, params={}):
         Resource.verify_keys(params, Customer.create_signature())
@@ -49,22 +44,6 @@ class CustomerGateway(object):
 
         response = self.config.http().post(self.config.base_merchant_path() + "/customers/advanced_search_ids", {"search": self.__criteria(query)})
         return ResourceCollection(query, response, self.__fetch)
-
-    def tr_data_for_create(self, tr_data, redirect_url):
-        Resource.verify_keys(tr_data, [{"customer": Customer.create_signature()}])
-        tr_data["kind"] = TransparentRedirect.Kind.CreateCustomer
-        return self.gateway.transparent_redirect.tr_data(tr_data, redirect_url)
-
-    def tr_data_for_update(self, tr_data, redirect_url):
-        Resource.verify_keys(tr_data, ["customer_id", {"customer": Customer.update_signature()}])
-        tr_data["kind"] = TransparentRedirect.Kind.UpdateCustomer
-        return self.gateway.transparent_redirect.tr_data(tr_data, redirect_url)
-
-    def transparent_redirect_create_url(self):
-        return self.config.base_url() + self.config.base_merchant_path() + "/customers/all/create_via_transparent_redirect_request"
-
-    def transparent_redirect_update_url(self):
-        return self.config.base_url() + self.config.base_merchant_path() + "/customers/all/update_via_transparent_redirect_request"
 
     def update(self, customer_id, params={}):
         Resource.verify_keys(params, Customer.update_signature())