+++ braintree/exceptions/service_unavailable_error.py
@@ -1,0 +1,7 @@
+from braintree.exceptions.braintree_error import BraintreeError
+
+class ServiceUnavailableError(BraintreeError):
+    """
+    Raised when the gateway is unavailable.
+    """
+    pass