+++ braintree/payment_method_nonce_gateway.py
@@ -12,9 +12,11 @@ class PaymentMethodNonceGateway(object):
         self.gateway = gateway
         self.config = gateway.config
 
-    def create(self, payment_method_token):
+    def create(self, payment_method_token, params = {"payment_method_nonce": {}}):
         try:
-            response = self.config.http().post(self.config.base_merchant_path() + "/payment_methods/" + payment_method_token + "/nonces")
+            schema = [{"payment_method_nonce": ["merchant_account_id", "authentication_insight", "amount"]}]
+            Resource.verify_keys(params, schema)
+            response = self.config.http().post(self.config.base_merchant_path() + "/payment_methods/" + payment_method_token + "/nonces", params)
             if "api_error_response" in response:
                 return ErrorResult(self.gateway, response["api_error_response"])
             else: