+++ braintree/webhook_testing_gateway.py
@@ -1,10 +1,7 @@
 from braintree.util.crypto import Crypto
 from braintree.webhook_notification import WebhookNotification
 import sys
-if sys.version_info[0] == 2:
-    from base64 import encodestring as encodebytes
-else:
-    from base64 import encodebytes
+from base64 import encodebytes
 from datetime import datetime
 
 class WebhookTestingGateway(object):
@@ -76,17 +73,6 @@ class WebhookTestingGateway(object):
             return self.__subscription_charged_unsuccessfully_sample_xml(id)
         elif kind == WebhookNotification.Kind.AccountUpdaterDailyReport:
             return self.__account_updater_daily_report_sample_xml()
-        # NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
-        # DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
-        elif kind == WebhookNotification.Kind.IdealPaymentComplete:
-            return self.__ideal_payment_complete_sample_xml(id)
-        # NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
-        # DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
-        elif kind == WebhookNotification.Kind.IdealPaymentFailed:
-            return self.__ideal_payment_failed_sample_xml(id)
-        # NEXT_MAJOR_VERSION remove GrantedPaymentInstrumentUpdate
-        elif kind == WebhookNotification.Kind.GrantedPaymentInstrumentUpdate:
-            return self.__granted_payment_instrument_update()
         elif kind == WebhookNotification.Kind.GrantorUpdatedGrantedPaymentMethod:
             return self.__granted_payment_instrument_update()
         elif kind == WebhookNotification.Kind.RecipientUpdatedGrantedPaymentMethod:
@@ -584,40 +570,6 @@ class WebhookTestingGateway(object):
                 <report-url>link-to-csv-report</report-url>
             </account-updater-daily-report>
             """
-
-    # NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
-    # DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
-    def __ideal_payment_complete_sample_xml(self, id):
-        return """
-            <ideal-payment>
-                <id>%s</id>
-                <status>COMPLETE</status>
-                <issuer>ABCISSUER</issuer>
-                <order-id>ORDERABC</order-id>
-                <currency>EUR</currency>
-                <amount>10.00</amount>
-                <created-at>2016-11-29T23:27:34.547Z</created-at>
-                <approval-url>https://example.com</approval-url>
-                <ideal-transaction-id>1234567890</ideal-transaction-id>
-            </ideal-payment>
-            """ % id
-
-    # NEXT_MAJOR_VERSION Remove this class as legacy Ideal has been removed/disabled in the Braintree Gateway
-    # DEPRECATED If you're looking to accept iDEAL as a payment method contact accounts@braintreepayments.com for a solution.
-    def __ideal_payment_failed_sample_xml(self, id):
-        return """
-            <ideal-payment>
-                <id>%s</id>
-                <status>FAILED</status>
-                <issuer>ABCISSUER</issuer>
-                <order-id>ORDERABC</order-id>
-                <currency>EUR</currency>
-                <amount>10.00</amount>
-                <created-at>2016-11-29T23:27:34.547Z</created-at>
-                <approval-url>https://example.com</approval-url>
-                <ideal-transaction-id>1234567890</ideal-transaction-id>
-            </ideal-payment>
-            """ % id
 
     def __granted_payment_instrument_update(self):
         return """