+++ tests/integration/test_merchant_account.py
@@ -1,29 +1,6 @@
 from tests.test_helper import *
 
 class TestMerchantAccount(unittest.TestCase):
-    DEPRECATED_APPLICATION_PARAMS = {
-        "applicant_details": {
-            "company_name": "Garbage Garage",
-            "first_name": "Joe",
-            "last_name": "Bloggs",
-            "email": "joe@bloggs.com",
-            "phone": "555-555-5555",
-            "address": {
-                "street_address": "123 Credibility St.",
-                "postal_code": "60606",
-                "locality": "Chicago",
-                "region": "IL",
-            },
-            "date_of_birth": "10/9/1980",
-            "ssn": "123-00-1234",
-            "tax_id": "123456789",
-            "routing_number": "122100024",
-            "account_number": "43759348798"
-        },
-        "tos_accepted": True,
-        "master_merchant_account_id": "sandbox_master_merchant_account"
-    }
-
     VALID_APPLICATION_PARAMS = {
         "individual": {
             "first_name": "Joe",
@@ -59,13 +36,6 @@ class TestMerchantAccount(unittest.TestCase):
         "tos_accepted": True,
         "master_merchant_account_id": "sandbox_master_merchant_account"
     }
-
-    def test_create_accepts_deprecated_parameters(self):
-        result = MerchantAccount.create(self.DEPRECATED_APPLICATION_PARAMS)
-
-        self.assertTrue(result.is_success)
-        self.assertEqual(MerchantAccount.Status.Pending, result.merchant_account.status)
-        self.assertEqual("sandbox_master_merchant_account", result.merchant_account.master_merchant_account.id)
 
     def test_create_application_with_valid_params_and_no_id(self):
         result = MerchantAccount.create(self.VALID_APPLICATION_PARAMS)