+++ tests/unit/test_payment_method_gateway.py
@@ -1,9 +1,6 @@
 from tests.test_helper import *
 from braintree.payment_method_gateway import PaymentMethodGateway
-if sys.version_info[0] == 2:
-    from mock import MagicMock
-else:
-    from unittest.mock import MagicMock
+from unittest.mock import MagicMock
 
 class TestPaymentMethodGateway(unittest.TestCase):
     def test_create_signature(self):
@@ -22,7 +19,6 @@ class TestPaymentMethodGateway(unittest.TestCase):
             "number",
             "payment_method_nonce",
             "paypal_refresh_token",
-            "paypal_vault_without_upgrade",
             "token",
             {
                 "billing_address": Address.create_signature()},